Conflicts:
	.gitignore
	salt/template.py
	salt/utils/__init__.py
	salt/utils/templates.py

+++ salt/modules/timezone.py
@@ -2,7 +2,10 @@
 Module for managing timezone on posix-like systems.
 '''
 
+# Import python libs
 import os
+import hashlib
+import salt.utils
 import logging
 
 log = logging.getLogger(__name__)
@@ -36,6 +39,11 @@ def get_zone():
         cmd = 'grep ZONE /etc/sysconfig/clock | grep -vE "^#"'
     elif 'Debian' in __grains__['os_family']:
         return open('/etc/timezone','r').read()
+    elif 'Gentoo' in __grains__['os_family']:
+        return open('/etc/timezone','r').read()
+    elif 'FreeBSD' in __grains__['os_family']:
+        return ('FreeBSD does not store a human-readable timezone. Please'
+                'consider using timezone.get_zonecode or timezone.zonecompare')
     out = __salt__['cmd.run'](cmd).split('=')
     ret = out[1].replace('"', '')
     return ret
@@ -69,7 +77,11 @@ def get_offset():
 
 def set_zone(timezone):
     '''
-    Unlinks, then symlinks /etc/localtime to the set timezone
+    Unlinks, then symlinks /etc/localtime to the set timezone.
+
+    The timezone is crucial to several system processes, each of which SHOULD
+    be restarted (for instance, whatever you system uses as its cron and
+    syslog daemons). This will not be magically done for you!
 
     CLI Example::
 
@@ -79,7 +91,8 @@ def set_zone(timezone):
     if not os.path.exists(zonepath):
         return 'Zone does not exist: {0}'.format(zonepath)
 
-    os.unlink('/etc/localtime')
+    if os.path.exists('/etc/localtime'):
+        os.unlink('/etc/localtime')
     os.symlink(zonepath, '/etc/localtime')
 
     if 'Arch' in __grains__['os_family']:
@@ -88,8 +101,36 @@ def set_zone(timezone):
         __salt__['file.sed']('/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
     elif 'Debian' in __grains__['os_family']:
         open('/etc/timezone', 'w').write(timezone)
+    elif 'Gentoo' in __grains__['os_family']:
+        open('/etc/timezone', 'w').write(timezone)
 
     return True
+
+
+def zone_compare(timezone):
+    '''
+    Checks the md5sum between the given timezone, and the one set in
+    /etc/localtime. Returns True if they match, and False if not. Mostly useful
+    for running state checks.
+
+    Example::
+
+        salt '*' timezone.zone_compare 'America/Denver'
+    '''
+    if not os.path.exists('/etc/localtime'):
+        return 'Error: /etc/localtime does not exist.'
+
+    zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
+
+    with salt.utils.fopen(zonepath, 'r') as fp_:
+        usrzone = hashlib.md5(fp_.read()).hexdigest()
+
+    with salt.utils.fopen('/etc/localtime', 'r') as fp_:
+        etczone = hashlib.md5(fp_.read()).hexdigest()
+
+    if usrzone == etczone:
+        return True
+    return False
 
 
 def get_hwclock():
@@ -115,6 +156,10 @@ def get_hwclock():
             return 'UTC'
         else:
             return 'localtime'
+    elif 'Gentoo' in __grains__['os_family']:
+        cmd = 'grep "^clock=" /etc/conf.d/hwclock | grep -vE "^#"'
+        out = __salt__['cmd.run'](cmd).split('=')
+        return out[1].replace('"', '')
 
 
 def set_hwclock(clock):
@@ -125,6 +170,7 @@ def set_hwclock(clock):
 
         salt '*' timezone.set_hwclock UTC
     '''
+    timezone = get_zone()
     zonepath = '/usr/share/zoneinfo/{0}'.format(timezone)
     if not os.path.exists(zonepath):
         return 'Zone does not exist: {0}'.format(zonepath)
@@ -133,7 +179,7 @@ def set_hwclock(clock):
     os.symlink(zonepath, '/etc/localtime')
 
     if 'Arch' in __grains__['os_family']:
-        __salt__['file.sed']('/etc/rc.conf', '^HARDWARECLOCK=.*', 'HARDWARECLOCK="{0}"'.format(timezone))
+        __salt__['file.sed']('/etc/rc.conf', '^HARDWARECLOCK=.*', 'HARDWARECLOCK="{0}"'.format(clock))
     elif 'RedHat' in __grains__['os_family']:
         __salt__['file.sed']('/etc/sysconfig/clock', '^ZONE=.*', 'ZONE="{0}"'.format(timezone))
     elif 'Debian' in __grains__['os_family']:
@@ -141,6 +187,8 @@ def set_hwclock(clock):
             __salt__['file.sed']('/etc/default/rcS', '^UTC=.*', 'UTC=yes')
         elif clock == 'localtime':
             __salt__['file.sed']('/etc/default/rcS', '^UTC=.*', 'UTC=no')
+    elif 'Gentoo' in __grains__['os_family']:
+        __salt__['file.sed']('/etc/conf.d/hwclock', '^clock=.*', 'clock="{0}"'.format(clock))
 
     return True
